Standardize RequiredComponent creation and deletion
The SliderJoint needs to be linked to a ConfigurableJoint in order to work properly. Ensuring a link can be quite challening because of external factors like the component resetting, the editor reloading or Undo operations being performed.

The RequiredComponent<> type has been fit to handle these scenarios in a way that requires minimum maintanance. One can simply adhere to this pattern if they want to link a component (if using the RequireComponentAttribute is not a proper solution).

```
[SerializeField][HideInInspector] private RequiredComponent<MyComponent> _myComponent;
public MyComponent myComponent=> _myComponent.GetComponent(gameObject);

private void OnDestroy()
{
    // Note: it is recommended to use ExtraObject.DestroySafe(_myComponent) if you plan on linking the component inside the editor with e.g. OnValidate or [ExecuteAlways]
    Destroy(_myComponent);
}
```

Making calls to myComponent will ensure a component. If it doesn't exist, it will be created immediately.