Add UnityExtras.InputSystem.Input
Input is a convenient type that allows you to add input in such a way that you can always change your setup later without changing your code. It consists of:
- An InputActionProperty, which allows you to choose between either an InputActionReference or a custom constructed InputAction.
- A bool to specify if you want to enable the action on start or not.
- A foldout to hide the action for a sleeker inspector.

The intended use case for Input is that you can start off with a custom InputAction and the enableOnStart toggle set to true. This allows you to test your input immediately.
Once you're satisfied, you may change to an InputActionReference from a dedicated InputActionAsset.
If later you implement ways for different input to be turned on or off, you may disable the enableOnStart toggle and let your own code take over.

Input has already been added to both the FirstPersonCharacter and Picker scripts, and should be used whenever a script needs to reference Input.