Overhaul Continuous Input Implementation
ContinuousInteraction was a complete failure. A new way of adding Continuous actions has been developed using references and extension methods, and it's dead-easy to use!

On any InputAction, simply call `AddContinuousActions(params Action<InputAction.CallbackContext>[])` or `RemoveContinuousActions(params Action<InputAction.CallbackContext>[])` and the action will be called continuously as long as the Input is performed!

Continuous Actions allow for code to remain simple and straightforward, with scripts that implement it not having to worry about caching a bunch of values to check if the InputAction is being performed or not while still using the Action pattern.

Additionally, Continuous Actions automatically provide accessability options! If you want an action to be called when an input is performed, do nothing. But if you want a single-fire action that cancels after a duration (useful for variable jumps), simply add a "Tap" interaction. For an action you want to be called continuously until you press the input again, add a "Press: Release Only" interaction. And for an action that cancels on release, add a "SlowTap" interaction. These can be combined and changed to allow players accessible input options without you having to provide any extra code for it!