Add IAuthor, my pièce de résistance
IAuthor is an interface that, together with RequiredComponent<T>, provides an effective way to Author components from your custom scripts. CharacterController2D and SliderJoint implement this to cleanly Author a CapsuleCollider2D and ConfigurableJoint respectively.

In the background, IAuthor handles some serialization for you to ensure data between 2 components can be neatly carried over inside the IAuthor.Serialize and IAuthor.Deserialize methods, as well as appropriately reset and destroyed from IAuthor.ResetAuthor and IAuthor.DestroyAuthor.

It is crucial that IAuthor be implemented properly, thus a code example has been provided in the comments of IAuthor. Use this as a starting point when in doubt about its proper implementation.