Rewrite CharacterMover and CharacterMover2D
CharacterMover and CharacterMover2D were sharing a lot of methods and values. In order to adhere to a DRY-er structure of code writing they now inherit from a base class CharacterMoverBase, which implements all these shared methods and allows for any changes to it to be reflected in both classes.

An "allowPerpetualJump" bool has been added. If this is turned off, the character can't jump perpetually and the Jump has to be released in order for the character to be able to jump again.

Additionally, a ForcedJump method has been added that ignores this check, as well as ignoring if the character is grounded or not. In this way, ForcedJump could potentially be used to allow double jumping.