Add Picker2D and PickUp2D, make base for 2D and 3D Picker and PickUp to inherit from
Picklings is a script containing all of the base methods that are expected of a Picker and PickUp. Picker and PickUp have been made as to inherit from those base classes, but should remain functionally untouched.

As well, a Picker2D and PickUp2D have been added which inherit from the same base but, of course, work in 2D.

Additionally, ValueStore has been given its own script for reuse elsewhere. ValueStore allows for a value to be stored and replaced temporarily. This is useful in PickUp as you want a rigidbody to have different settings during its PickUp state (e.g. gravity = false) but reset this value once you get out of this state.