Change SerializedPropertyExtensions.GetMember implementation
SerializedPropertyExtensions.GetMember() has several issues, most notably it would not correctly work with SetValue as that required all parent SerializedProperties to be set as well, else the value would reset to what was serialized.

The new class PropertyMember is intended to hold a SerializedProperty's MemberInfo, as well as other data that it needs to get and set values correctly. This implementation subsequently allows you to cache this information, much like GetField and GetProperty allowed you to do.

Since this new implementation fixes the old one, the old implementation has been removed completely. Additionally, ObjectDrawer and ObjectPicker use PropertyMember to get the SerializedPropertyType: this standardizes the implementation resulting in less maintenance.