Add LinkPropertyAttribute
LinkProperty allows you to link a unity field to a property.

Example:
```csharp
[SerializeField, LinkProperty(nameof(value))] private float _value; // Will be implicitly clamped to 0 - 1 by the property.
public float value
{
    get => _value;
    set => _value = Mathf.Clamp01(value);
}
```

This may be used to standardize code by writing properties that correctly modify values and inspector fields that correctly display the modified versions.

Note that in some cases, the DelayAttribute must be added in order for the inspector to allow modification as expected.