Develop (#5)
* Move Menus into "Menus" folder

Also rename "ContextualMenus" script to "Menus.Contextual" and use the partial class for it.

* Add RotateAnimationsToMovePerfectlyForwards menu

This will rotate animations on model importers to move perfectly forwards (duh)

* Fix an issue in LinkPropertyDrawer

It would not always update correctly if a property changed through another property, but that has been fixed.

* Add AnimatorParameters

These structs allow for linking up to Animator Parameters in a clean manner where all settings are kept in one place and easy to code for. They also have nice editors allowing you to test different values for your animations inside the inspector.

* Moved the Rigging folder under the Animation folder

* Remove the AnimatorHash.uxml

It was excessive, so I opted to simply bind the property in script. This makes it easier to share AnimatorHashDrawer with friends.