Fix AnimatorParameter Issues and improve ease-of-use
The AnimatorParameters would not display correctly in the inspector, with the foldout not being clickable and the AnimatorHash being drawn in front of it. These display issues have been fixed.

Additionally, all AnimatorParameters are classes now as structs would trigger Error Code CS1612 "Cannot modify the return value of 'expression' because it is not a variable". Simply copying the value and placing it back would not be a great option since the current architecture relies on a persistent data model because of its ties with Action onValueChanged.

Speaking of which, onValueChanged will now trigger even if the new value is the same as the old value. This is to allow for animator parameters to be changed by different means than the Parameter class itself.

And lastly, a default coroutine has been added under the AnimatorExtensions to allow for quickly setting up coroutines with FloatParameters as these often have a damptime.