Add ScriptableAnimationEvent
AnimationEvents have a very rigid architecture, calling methods by name and only allowing for a handful of limited forms of data to be attached to it.

ScriptableAnimationEvents, while not a perfect system, are a very easy way to use AnimationEvents in a modular fashion. All that is required is to attach a ScriptableAnimationEventListener to your Animator, to name all your functions "Play", and to attach a ScriptableAnimationEvent as object data. The ScriptableAnimationEvent will then play as intended.

ScriptableAnimationEvents are ScriptableObjects, so they can be created as data packets. It's core functionality is that a ScriptableAnimationEvent can have children of ScriptableAnimationEvents, meaning you can easily link different events together and mix them for different situations.