Add CharacterSlopeSlider
This component makes CharacterControllers slide off of slopes. Normally the CharacterController reaching its slope limit stands still. This is a problem when your character can jump and move in the air, as it could simply jump up any slope without issue. This component slides a character down a slope and prevents it from jumping.