Add Instance<T> and CursorInstance
Instance is a simple way to set static in Unity via components. Singletons carry many interdependencies with them, but Instances operate completely autonomously. The current instance is always the one that was added last. It provides a SetState method that is called on enable. When the component is disabled, the last known instance will be enabled.

CursorInstance has been provided as an example of how to use Instance. When set state gets called, it will simply set the state of the static Unity cursor.