Change version to 1.2.1 (#8)
* Add TimeInstance

This is an instance to set TimeScale via Instances.

* Add First Person Character Sample

This lets you quickly import a first person character in your scene to experiment with.

* Add Log methods to MonoEvent

Remove MonoEvent's debug messages, that would trigger an unnecessary if check. Instead, it now provides a log method that you can link inside your event in order to log messages when you feel it is appropriate.

* Improve Switches

Switches have been reworked to work more logically and allow for more use cases.

- `reverse` has been renamed to `invertSwitch`, which is much more explicit.
- `isOn` has been made public and takes `invertSwitch` into account much more appropriately. e.g. if you call `isOn = false` when `invertSwitch == true`, the `on` event will be called as expected.
- `ConditionalSwitch` is an extension of `Switch` that replaces the old switch behaviour. `ConditionalSwitch` has an `Update` call, `Switch` does not, meaning it can be used with less overhead.
- `LocalPositionSwitch` has been renamed to `PositionSwitch` and inherits from `ConditionalSwitch`.
- `PositionSwitch`, instead of taking a `Vector3 targetLocalPosition`, now takes a `Transform target` and a `Vector3 offset`. This provides much more flexibility where the switch can easily be made local by providing the target with the parent of the object, while also allowing for more use cases where the target can move around or the offset be set in world space.

* Add InputSwitch

`InputSwitch` inherit from the improved `Switch` architecture in order to invoke `on` and `off` via an input button, without incurring any overhead on update.

* Change version to 1.2.1

Change the package version and the changelog to accomodate.